+++ src/components/Cards/index.ts
@@ -3,4 +3,4 @@ export { default as EventsCard } from './events-card'
 export { default as LinkCard } from './link-card'
 export { default as LoginCard } from './login-card'
 export { default as SportsCard } from './sports-card'
-export { default as UpNextCard } from './up-next-card'
+export { UpNextCard } from './up-next'
