+++ src/components/Cards/up-next/index.ts
@@ -1,0 +1,1 @@
+export { default as UpNextCard } from './up-next-card'
